Fix editor structs (#9286)
