+++ src/common/structs/is-icon.ts
@@ -1,10 +1,5 @@
 import { refine, string } from "superstruct";
 
-const isIcon = (value: string) => {
-  if (!value.includes(":")) {
-    return false;
-  }
-  return true;
-};
+const isIcon = (value: string) => value.includes(":");
 
 export const icon = () => refine(string(), "icon (mdi:icon-name)", isIcon);
