+++ src/panels/lovelace/editor/config-elements/hui-generic-entity-row-editor.ts
@@ -14,11 +14,8 @@ import "../../components/hui-action-editor";
 import "../../components/hui-entity-editor";
 import "../../components/hui-theme-select-editor";
 import { LovelaceRowEditor } from "../../types";
-import {
-  EditorTarget,
-  entitiesConfigStruct,
-  EntitiesEditorEvent,
-} from "../types";
+import { entitiesConfigStruct } from "../structs/entities-struct";
+import { EditorTarget, EntitiesEditorEvent } from "../types";
 import { configElementStyle } from "./config-elements-style";
 
 const SecondaryInfoValues: { [key: string]: { domains?: string[] } } = {
